+++ lib/src/core/models.dart
@@ -1,9 +1,6 @@
-library subtitle.core;
-
 /// This model class store the subtitle data.
 class Subtitle {
-  /// Current index of this subtitle in its file.
-  /// The **-1** value mean that is no index for this subtitle in its file.
+  /// Current index number of this subtitle in its file.
   final int index;
 
   /// Store the current text for periode that started with [start] and
@@ -34,6 +31,10 @@ class Subtitle {
   int compareTo(Subtitle other) =>
       start.inMilliseconds.compareTo(other.start.inMilliseconds);
 
+  bool inRange(Duration duration) => start <= duration && end >= duration;
+  bool isLarg(Duration duration) => duration > end;
+  bool inSmall(Duration duration) => duration < start;
+
   @override
   bool operator ==(Object other) {
     if (other is Subtitle) {
@@ -50,6 +51,9 @@ class Subtitle {
 
   @override
   int get hashCode => props.hashCode;
+
+  @override
+  String toString() => '$start --> $end\n$data';
 
   List<Object> get props => [start, end, data, index];
 }