+++ lib/subtitle.dart
@@ -1,10 +1,11 @@
 library subtitle;
 
-export 'core/models.dart';
-export 'core/exceptions.dart';
+export 'src/core/models.dart';
+export 'src/core/exceptions.dart';
 
-export 'utils/types.dart';
-export 'utils/regexes.dart';
-export 'utils/subtitle_parser.dart';
-export 'utils/subtitle_provider.dart';
-export 'utils/subtitle_repository.dart';
+export 'src/utils/types.dart';
+export 'src/utils/regexes.dart';
+export 'src/utils/subtitle_parser.dart';
+export 'src/utils/subtitle_provider.dart';
+export 'src/utils/subtitle_repository.dart';
+export 'src/utils/subtitle_controller.dart';